C++: all: Makefile: align compiler flags
Some C++ examples could not be cross-compiled for RPi using export
CXX=arm-linux-gnueabihf-g++  because some compiler variables could not
be set from outside or were compiled with C compiler. After this change,
all examples will be compiled with C++ compiler that can be overridden
with an environment variable set to the cross-compiler.
This commit doesn't make everything cross-compile as there's a
dependency on pigpio library that should be addressed in other commits.